Publish target, action, and failure information
BuildBuddy uses pattern expand messages to collect the set of targets to report on.
Buck2 does not have a direct correspondance to the target completed message, for now we collect all
targets for which actions are completed and then emit a series of target completed events before we
close the stream.
Stdout and stderr are only included inline in Buck2 and not as CAS items, so we forward them as
inline data.
Failure details are used by BuildBuddy to display build errors in the UI, so we emit these.